Add ability to mount volumes from host
When a volume is mounted using the -v command, the contents of the
volumes are erased, which removes all of the dokuwiki files needed
to get things working. As a working around, the volumes are copied
to another folder, and are re-copied (without overwriting) back to
their original location once the image is running.